decals/requirements.txt
PyOpenGL==3.1.0
pywavefront==1.3.0
numpy==1.17.4
noise==1.2.2